databricks-genie/genie-gradio-app/requirements.txt
gradio==4.44.1
databricks-sql-connector==3.3.0
python-dotenv==1.0.0